refactor: make utility a common file extension
