+++ src/utility/index.ts
@@ -1,7 +1,7 @@
 import { Config } from '../config/index';
-import { WebStorageUtility } from './webstorage-utility';
+import { WebStorageUtility } from './webstorage.utility';
 import { cookiesStorage } from './storage/cookies-storage';
-import { SharedStorageUtility } from './shared-utility';
+import { SharedStorageUtility } from './shared-storage.utility';
 import { sharedStorage } from './storage/shared-storage';
 
 export const localStorageUtility: WebStorageUtility =
